+++ db/migrate/20220906094506_add_caseworker_assigned_at_to_project.rb
@@ -1,0 +1,5 @@
+class AddCaseworkerAssignedAtToProject < ActiveRecord::Migration[7.0]
+  def change
+    add_column :projects, :caseworker_assigned_at, :datetime
+  end
+end