+++ spec/models/project_spec.rb
@@ -7,6 +7,7 @@ RSpec.describe Project, type: :model do
     it { is_expected.to have_db_column(:target_completion_date).of_type :date }
     it { is_expected.to have_db_column(:caseworker_id).of_type :uuid }
     it { is_expected.to have_db_column(:team_leader_id).of_type :uuid }
+    it { is_expected.to have_db_column(:caseworker_assigned_at).of_type :datetime }
   end
 
   describe "Relationships" do